+++ json_from_csv/handler.py
@@ -1,10 +1,7 @@
-import os
 import boto3
 import json
-import sys
-sys.path.append(os.path.dirname(os.path.realpath(__file__)))
 from processCsv import processCsv
-
+from pathlib import Path
 
 def run(event, context):
     id = event.get("id")
@@ -39,16 +36,18 @@ def write_s3_json(s3Bucket, s3Path, json_hash):
 
 # python -c 'from handler import *; test()'
 def test():
-    with open("../example/item-one-image/config.json", 'r') as input_source:
+    current_path = Path(__file__).absolute()
+
+    with open(current_path + "/../example/item-one-image/config.json", 'r') as input_source:
         config = json.load(input_source)
     input_source.close()
-    with open("../example/item-one-image/main.csv", 'r') as input_source:
+    with open(current_path + "/../example/item-one-image/main.csv", 'r') as input_source:
         main_csv = input_source.read()
     input_source.close()
-    with open("../example/item-one-image/items.csv", 'r') as input_source:
+    with open(current_path + "/../example/item-one-image/items.csv", 'r') as input_source:
         items_csv = input_source.read()
     input_source.close()
-    with open("../example/item-one-image/image-data.json", 'r') as input_source:
+    with open(current_path + "/../example/item-one-image/image-data.json", 'r') as input_source:
         image_data = json.load(input_source)
     input_source.close()
