--- manifest-from-input-json/test_iiifManifest_unittest.py
@@ -1,42 +1,0 @@
-import unittest
-import json
-from iiifManifest import iiifManifest
-
-
-class TestIiifManifest(unittest.TestCase):
-    def setUp(self):
-        self.parent_id = 'parent_id'
-
-        with open("../example/example-input.json", 'r') as input_source:
-            self.example_data = json.load(input_source)
-        input_source.close()
-
-        self.iiifManifest = iiifManifest(self.parent_id, self.example_data.get('config'), self.example_data)
-        pass
-
-    def test_id(self):
-        self.assertEqual(self.iiifManifest.id, self.parent_id)
-
-    def test_config(self):
-        self.assertDictEqual(self.iiifManifest.config, self.example_data.get('config'))
-
-    def test_example_data(self):
-        self.assertDictEqual(self.iiifManifest.manifest_data, self.example_data)
-
-    def test_thumbnail(self):
-        test = {'@id': 'https://image-server.library.nd.edu:8182/iiif/2/parent_id%2F009_output/full/250,/0/default.jpg', 'service': {'@id': 'https://image-server.library.nd.edu:8182/iiif/2/parent_id%2F009_output', 'profile': 'http://iiif.io/api/image/2/level2.json', '@context': 'http://iiif.io/api/image/2/context.json'}}
-        self.assertDictEqual(self.iiifManifest.thumbnail(), test)
-
-    def test_sequences(self):
-        test = [{'@id': 'https://manifest.nd.edu//parent_id/sequence/1', '@type': 'sc:Sequence', 'label': 'sequency', 'viewingHint': 'individuals', 'canvases': [{'@id': 'https://manifest.nd.edu//parent_id/canvas/parent_id%2F009_output', '@type': 'sc:Canvas', 'label': '009', 'height': 2000, 'width': 2000, 'images': [{'@id': 'https://image-server.library.nd.edu:8182/iiif/2/parent_id%2F009_output', '@type': 'oa:Annotation', 'motivation': 'sc:painting', 'on': 'https://manifest.nd.edu//parent_id/canvas/parent_id%2F009_output', 'resource': {'@id': 'https://image-server.library.nd.edu:8182/iiif/2/parent_id%2F009_output/full/full/0/default.jpg', '@type': 'dctypes:Image', 'format': 'image/jpeg', 'service': {'@id': 'https://image-server.library.nd.edu:8182/iiif/2/parent_id%2F009_output', 'profile': 'http://iiif.io/api/image/2/level2.json', '@context': 'http://iiif.io/api/image/2/context.json'}}}], 'thumbnail': {'@id': 'https://image-server.library.nd.edu:8182/iiif/2/parent_id%2F009_output/full/250,/0/default.jpg', 'service': {'@id': 'https://image-server.library.nd.edu:8182/iiif/2/parent_id%2F009_output', 'profile': 'http://iiif.io/api/image/2/level2.json', '@context': 'http://iiif.io/api/image/2/context.json'}}}, {'@id': 'https://manifest.nd.edu//parent_id/canvas/parent_id%2F046_output', '@type': 'sc:Canvas', 'label': '046', 'height': 2000, 'width': 2000, 'images': [{'@id': 'https://image-server.library.nd.edu:8182/iiif/2/parent_id%2F046_output', '@type': 'oa:Annotation', 'motivation': 'sc:painting', 'on': 'https://manifest.nd.edu//parent_id/canvas/parent_id%2F046_output', 'resource': {'@id': 'https://image-server.library.nd.edu:8182/iiif/2/parent_id%2F046_output/full/full/0/default.jpg', '@type': 'dctypes:Image', 'format': 'image/jpeg', 'service': {'@id': 'https://image-server.library.nd.edu:8182/iiif/2/parent_id%2F046_output', 'profile': 'http://iiif.io/api/image/2/level2.json', '@context': 'http://iiif.io/api/image/2/context.json'}}}], 'thumbnail': {'@id': 'https://image-server.library.nd.edu:8182/iiif/2/parent_id%2F046_output/full/250,/0/default.jpg', 'service': {'@id': 'https://image-server.library.nd.edu:8182/iiif/2/parent_id%2F046_output', 'profile': 'http://iiif.io/api/image/2/level2.json', '@context': 'http://iiif.io/api/image/2/context.json'}}}, {'@id': 'https://manifest.nd.edu//parent_id/canvas/parent_id%2F2018_009', '@type': 'sc:Canvas', 'label': '2018 009', 'height': 2000, 'width': 2000, 'images': [{'@id': 'https://image-server.library.nd.edu:8182/iiif/2/parent_id%2F2018_009', '@type': 'oa:Annotation', 'motivation': 'sc:painting', 'on': 'https://manifest.nd.edu//parent_id/canvas/parent_id%2F2018_009', 'resource': {'@id': 'https://image-server.library.nd.edu:8182/iiif/2/parent_id%2F2018_009/full/full/0/default.jpg', '@type': 'dctypes:Image', 'format': 'image/jpeg', 'service': {'@id': 'https://image-server.library.nd.edu:8182/iiif/2/parent_id%2F2018_009', 'profile': 'http://iiif.io/api/image/2/level2.json', '@context': 'http://iiif.io/api/image/2/context.json'}}}], 'thumbnail': {'@id': 'https://image-server.library.nd.edu:8182/iiif/2/parent_id%2F2018_009/full/250,/0/default.jpg', 'service': {'@id': 'https://image-server.library.nd.edu:8182/iiif/2/parent_id%2F2018_009', 'profile': 'http://iiif.io/api/image/2/level2.json', '@context': 'http://iiif.io/api/image/2/context.json'}}}, {'@id': 'https://manifest.nd.edu//parent_id/canvas/parent_id%2F2018_049_004', '@type': 'sc:Canvas', 'label': '2018 049 004', 'height': 2000, 'width': 2000, 'images': [{'@id': 'https://image-server.library.nd.edu:8182/iiif/2/parent_id%2F2018_049_004', '@type': 'oa:Annotation', 'motivation': 'sc:painting', 'on': 'https://manifest.nd.edu//parent_id/canvas/parent_id%2F2018_049_004', 'resource': {'@id': 'https://image-server.library.nd.edu:8182/iiif/2/parent_id%2F2018_049_004/full/full/0/default.jpg', '@type': 'dctypes:Image', 'format': 'image/jpeg', 'service': {'@id': 'https://image-server.library.nd.edu:8182/iiif/2/parent_id%2F2018_049_004', 'profile': 'http://iiif.io/api/image/2/level2.json', '@context': 'http://iiif.io/api/image/2/context.json'}}}], 'thumbnail': {'@id': 'https://image-server.library.nd.edu:8182/iiif/2/parent_id%2F2018_049_004/full/250,/0/default.jpg', 'service': {'@id': 'https://image-server.library.nd.edu:8182/iiif/2/parent_id%2F2018_049_004', 'profile': 'http://iiif.io/api/image/2/level2.json', '@context': 'http://iiif.io/api/image/2/context.json'}}}]}]
-
-        self.assertListEqual(self.iiifManifest.sequences(), test)
-
-    def test_manifest(self):
-        test = {'@context': 'http://iiif.io/api/presentation/2/context.json', '@type': 'sc:Manifest', '@id': 'https://manifest.nd.edu//parent_id/manifest', 'label': 'Label', 'metadata': [{'label': 'Title', 'value': 'Wunder der Vererbung'}, {'label': 'Author(s)', 'value': 'Bolle, Fritz'}, {'label': 'Publication date', 'value': '[1951]'}, {'label': 'Attribution', 'value': 'Wellcome Library<br/>License: CC-BY-NC'}], 'description': 'Description', 'license': 'license info', 'attribution': 'attribution', 'viewingDirection': 'left-to-right', 'thumbnail': {'@id': 'https://image-server.library.nd.edu:8182/iiif/2/parent_id%2F009_output/full/250,/0/default.jpg', 'service': {'@id': 'https://image-server.library.nd.edu:8182/iiif/2/parent_id%2F009_output', 'profile': 'http://iiif.io/api/image/2/level2.json', '@context': 'http://iiif.io/api/image/2/context.json'}}, 'sequences': [{'@id': 'https://manifest.nd.edu//parent_id/sequence/1', '@type': 'sc:Sequence', 'label': 'sequency', 'viewingHint': 'individuals', 'canvases': [{'@id': 'https://manifest.nd.edu//parent_id/canvas/parent_id%2F009_output', '@type': 'sc:Canvas', 'label': '009', 'height': 2000, 'width': 2000, 'images': [{'@id': 'https://image-server.library.nd.edu:8182/iiif/2/parent_id%2F009_output', '@type': 'oa:Annotation', 'motivation': 'sc:painting', 'on': 'https://manifest.nd.edu//parent_id/canvas/parent_id%2F009_output', 'resource': {'@id': 'https://image-server.library.nd.edu:8182/iiif/2/parent_id%2F009_output/full/full/0/default.jpg', '@type': 'dctypes:Image', 'format': 'image/jpeg', 'service': {'@id': 'https://image-server.library.nd.edu:8182/iiif/2/parent_id%2F009_output', 'profile': 'http://iiif.io/api/image/2/level2.json', '@context': 'http://iiif.io/api/image/2/context.json'}}}], 'thumbnail': {'@id': 'https://image-server.library.nd.edu:8182/iiif/2/parent_id%2F009_output/full/250,/0/default.jpg', 'service': {'@id': 'https://image-server.library.nd.edu:8182/iiif/2/parent_id%2F009_output', 'profile': 'http://iiif.io/api/image/2/level2.json', '@context': 'http://iiif.io/api/image/2/context.json'}}}, {'@id': 'https://manifest.nd.edu//parent_id/canvas/parent_id%2F046_output', '@type': 'sc:Canvas', 'label': '046', 'height': 2000, 'width': 2000, 'images': [{'@id': 'https://image-server.library.nd.edu:8182/iiif/2/parent_id%2F046_output', '@type': 'oa:Annotation', 'motivation': 'sc:painting', 'on': 'https://manifest.nd.edu//parent_id/canvas/parent_id%2F046_output', 'resource': {'@id': 'https://image-server.library.nd.edu:8182/iiif/2/parent_id%2F046_output/full/full/0/default.jpg', '@type': 'dctypes:Image', 'format': 'image/jpeg', 'service': {'@id': 'https://image-server.library.nd.edu:8182/iiif/2/parent_id%2F046_output', 'profile': 'http://iiif.io/api/image/2/level2.json', '@context': 'http://iiif.io/api/image/2/context.json'}}}], 'thumbnail': {'@id': 'https://image-server.library.nd.edu:8182/iiif/2/parent_id%2F046_output/full/250,/0/default.jpg', 'service': {'@id': 'https://image-server.library.nd.edu:8182/iiif/2/parent_id%2F046_output', 'profile': 'http://iiif.io/api/image/2/level2.json', '@context': 'http://iiif.io/api/image/2/context.json'}}}, {'@id': 'https://manifest.nd.edu//parent_id/canvas/parent_id%2F2018_009', '@type': 'sc:Canvas', 'label': '2018 009', 'height': 2000, 'width': 2000, 'images': [{'@id': 'https://image-server.library.nd.edu:8182/iiif/2/parent_id%2F2018_009', '@type': 'oa:Annotation', 'motivation': 'sc:painting', 'on': 'https://manifest.nd.edu//parent_id/canvas/parent_id%2F2018_009', 'resource': {'@id': 'https://image-server.library.nd.edu:8182/iiif/2/parent_id%2F2018_009/full/full/0/default.jpg', '@type': 'dctypes:Image', 'format': 'image/jpeg', 'service': {'@id': 'https://image-server.library.nd.edu:8182/iiif/2/parent_id%2F2018_009', 'profile': 'http://iiif.io/api/image/2/level2.json', '@context': 'http://iiif.io/api/image/2/context.json'}}}], 'thumbnail': {'@id': 'https://image-server.library.nd.edu:8182/iiif/2/parent_id%2F2018_009/full/250,/0/default.jpg', 'service': {'@id': 'https://image-server.library.nd.edu:8182/iiif/2/parent_id%2F2018_009', 'profile': 'http://iiif.io/api/image/2/level2.json', '@context': 'http://iiif.io/api/image/2/context.json'}}}, {'@id': 'https://manifest.nd.edu//parent_id/canvas/parent_id%2F2018_049_004', '@type': 'sc:Canvas', 'label': '2018 049 004', 'height': 2000, 'width': 2000, 'images': [{'@id': 'https://image-server.library.nd.edu:8182/iiif/2/parent_id%2F2018_049_004', '@type': 'oa:Annotation', 'motivation': 'sc:painting', 'on': 'https://manifest.nd.edu//parent_id/canvas/parent_id%2F2018_049_004', 'resource': {'@id': 'https://image-server.library.nd.edu:8182/iiif/2/parent_id%2F2018_049_004/full/full/0/default.jpg', '@type': 'dctypes:Image', 'format': 'image/jpeg', 'service': {'@id': 'https://image-server.library.nd.edu:8182/iiif/2/parent_id%2F2018_049_004', 'profile': 'http://iiif.io/api/image/2/level2.json', '@context': 'http://iiif.io/api/image/2/context.json'}}}], 'thumbnail': {'@id': 'https://image-server.library.nd.edu:8182/iiif/2/parent_id%2F2018_049_004/full/250,/0/default.jpg', 'service': {'@id': 'https://image-server.library.nd.edu:8182/iiif/2/parent_id%2F2018_049_004', 'profile': 'http://iiif.io/api/image/2/level2.json', '@context': 'http://iiif.io/api/image/2/context.json'}}}]}]}
-
-        self.assertDictEqual(self.iiifManifest.manifest(), test)
-
-
-if __name__ == '__main__':
-    unittest.main()
