+++ manifest_from_input_json/handler.py
