+++ manifest_from_input_json/iiifImage.py
@@ -1,4 +1,4 @@
-from iiifItem import iiifItem
+from manifest_from_input_json.iiifItem import iiifItem
 
 
 class iiifImage(iiifItem):
