+++ manifest_from_input_json/iiifItem.py
